Increase upper bound on interfaces to 128
We hit the limit of 64 interfaces when running a large number of
`docker` containers, each of which created a bridge interface.  This
doubles the limit as a temporary work-around, although a real fix
would be to change the code to handle an arbitrarily large number of
interfaces as memory permits.